requirements.txt
opencv-python==4.13.0.90
pillow==12.1.0
